+++ test/jest/options/options.el
@@ -1,0 +1,37 @@
+;;; options.el --- Test all options flag  -*- lexical-binding: t; -*-
+
+;; Copyright (C) 2025  the Eask authors.
+
+;; Author: Shen, Jen-Chieh <jcs090218@gmail.com>
+;; Maintainer: Shen, Jen-Chieh <jcs090218@gmail.com>
+;; URL: https://github.com/emacs-eask/options
+;; Version: 0.0.0
+;; Package-Requires: ((emacs "26.1"))
+;; Keywords:  test
+
+;; This file is not part of GNU Emacs.
+
+;; This program is free software: you can redistribute it and/or modify
+;; it under the terms of the GNU General Public License as published by
+;; the Free Software Foundation, either version 3 of the License, or
+;; (at your option) any later version.
+
+;; This program is distributed in the hope that it will be useful,
+;; but WITHOUT ANY WARRANTY; without even the implied warranty of
+;; MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+;; GNU General Public License for more details.
+
+;; You should have received a copy of the GNU General Public License
+;; along with this program. If not, see <https://www.gnu.org/licenses/>.
+
+;;; Commentary:
+;;
+;; Test options.
+;;
+
+;;; Code:
+
+;; Empty..
+
+(provide 'options)
+;;; options.el ends here